[Liverpool] Make sure site-logo is always hidden on Desktop version
This logo is only used on the mobile version, on the Desktop version we
are using Liverpool's own header and logo. This fix ensures that the
logo never appears on the Desktop version, which was causing an issue
because some of the javascript causes the site logo to be displayed.