[Zurich] Refactor admin report print styles 677
Involves a not-as-bad-as-it-looks refactoring of the HTML markup for
report_edit.html and report_edit-sdm.html, with some actual indentation
so you can tell what's going on, and definition lists to give the
label/content pairs more logical structure.

Also replaces .report-edit-action with .print-only and .screen-only
which are less semantic, but much more obvious to newcomers.

fixup! [Zurich] Refactor admin report print styles 677

Allows SDMs to enlarge images on report edit page 718

fixup! [Zurich] Refactor admin report print styles 677

Displays category in report edit form on screen as well as in print